feat(captions): Allow multi-line captions using double newline separator
This change modifies how captions are parsed to allow for multi-line captions for each group. The separator for caption groups is changed from a single newline to two or more newlines.

This is achieved by using a regular expression to split the caption text, providing more flexibility in writing captions.